reshape the project structure
To avoid header conflict with users' existing stuff, I moved the
base class to a subdirectory. Now this would have not been an issue
internatlly, however, when the users had to include something, they
should have done
\#include <xQP/BaseTask>
if they install it while internally we were doing
\#include <BaseTask>
With the new changes, now everyone has to do the latter. It comes at
a cost of adding a new relatively empty directory but this is imposed
by cmake limits...